Fix long term storage with insurance plans
On startup, the client data was not loading the insurance plans
correctly from JSON saved data. As such it was possible to add plans
that a client already owns after closing and restarting the app.

This makes the app's usefulness for the user a lot worse as they have to
keep the app open all the time.

Let's fix this behaviour so the app can be loaded properly from storage.

This will make the app more useful to the user even if they closed the
app or the app crashes halfway.